Section 1 Day Example
There will be available 40 different sections in the Employee Scheduler. This is only 1 Section as an example. 

The earliest time an employee can be scheduled can be different than when the store is opened as either earlier or later. The number of people needed at this time is put as a requirement for the scheduler to fill out. 

It then goes on to ask how many people are needed at store opening. If the earliest time is later than store opening, the manager or trusted administrator is allowed to choose 0. Otherwise the number of people selected here will be as part of a requirement for the scheduler to meet.

The number of positions at the busiest time of day: This will be what makes the number of columns the scheduler must fill out. As the note says the scheduler will be able to blackout any area that is not necessary but will be required to fill in at least one shift.

Number of employees until store closes is self explanatory.

Latest time employees can be scheduled. This can be different than when stores close. 

The number of employees until latest time will be the employee requirement to the latest time available.