Employee Info Day
The Employee Info Day will be used to gather all employee for Day Shift schedule. It gathers if the employee is full time or part time and the time the employee is able to work. The time available is from Midnight to Midnight plus 5 hours. This information must be entered by a manager or trusted administrator. The manager or trusted administrator also has the ability to add important notes. If a note is added by the manager or administrator it will in appear in a pop-up box for the scheduler any time the employee is selected.

The Scheduling portion allows the scheduler to input full or partial day off requests. The Employee Scheduler will read these times and not allow the employee to be scheduled during this time. The scheduler is allowed to enter employee notes on why the manager notes are not being followed.

Where manager notes will be permanent, the scheduler notes will only be kept for only the time of current scheduling.