Scheduling
Here is where the scheduler will complete the scheduling for the period selected. The first day will come from the storeinfo provided by the Manager. The times the schedule will show will be coming from storeinfo and section1 (respectively). 

The employee selection will come from the employeeinfoDay or employeeinfoNight. They will only be available in the respective section selected and the time they are available. 

I am stuck on this part. I don't know how to pull this information to use here in a calendar format.